Move Exception to shared partition (#23361)
* Move Exception to shared partition

* Remove class name caching

* Not useful comments removal
